+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/DialogDeleteAllTexts.tsx
@@ -4,7 +4,7 @@ import { useParams } from "react-router-dom";
 import { useContextText } from "modules/texts/Text/Context";
 import { useDialogText } from "modules/texts/Text/shared/hooks/useDialogText";
 import DialogDeleteAllTextView from "./DialogDeleteAllTextsView";
-import { fetchDeleteAllTextsService } from "modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/service";
+import { fetchDeleteAllTextsService } from "modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/services";
 import { TEXT_QUERY_KEY } from "shared/services/constants/texts";
 
 const DialogDeleteAllTexts: React.FC = () => {
