+++ src/modules/texts/Text/components/Dialogs/DialogDeleteAllTexts/services/fetchDeleteAllTextsService.ts
